chatbot/dialogflow/requirements.txt
appdirs==1.4.3
APScheduler==3.6.3
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
Cython==0.29.22
distlib==0.3.0
distro==1.4.0
google-api-core==1.26.2
google-auth==1.28.0
google-cloud-dialogflow==2.0.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
ocrspace==2.2.1
packaging==20.3
pdf2image==1.14.0
pep517==0.8.2
Pillow==8.1.2
progress==1.5
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
PyPDF2==1.26.0
pytesseract==0.3.7
python-telegram-bot==13.4.1
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
retrying==1.3.3
rsa==4.7.2
six==1.14.0
tornado==6.1
tqdm==4.59.0
tzlocal==2.1
urllib3==1.25.8
webencodings==0.5.1
